+++ packages/js/src/plugins/collateralModule/operations/findCollateralByAddress.ts
@@ -1,6 +1,8 @@
 import { PublicKey } from '@solana/web3.js';
 import { Collateral, toCollateral } from '../models';
 import { toCollateralAccount } from '../accounts';
+import { toRegisteredMint } from '@/plugins/protocolModule';
+import { toRegisteredMintAccount } from '@/plugins/protocolModule/accounts';
 import {
   Operation,
   OperationHandler,
@@ -74,13 +76,22 @@ export const findCollateralByAddressOperationHandler: OperationHandler<FindColla
       // const collateralMint = await convergence
       //   .protocol()
       //   .findRegisteredMintByAddress({ address: protocol.collateralMint });
+      //@ts-ignore
       const collateralMint = await convergence
         .tokens()
         .findMintByAddress({ address: protocol.collateralMint });
 
+      const mintInfo = convergence
+        .rfqs()
+        .pdas()
+        .mintInfo({ mint: protocol.collateralMint });
+      const acct = await convergence.rpc().getAccount(mintInfo);
+      const registeredMint = toRegisteredMint(toRegisteredMintAccount(acct));
+
       collateralModel.lockedTokensAmount /= Math.pow(
         10,
-        collateralMint.decimals
+        // collateralMint.decimals
+        registeredMint.decimals
       );
 
       return collateralModel;