+++ packages/js/src/plugins/rpcModule/RpcClient.ts
@@ -111,6 +111,10 @@ export class RpcClient {
   ): Promise<Transaction> {
     const { keypairs, identities } = getSignerHistogram(signers);
 
+    const recentBlockhash =
+      await this.convergence.connection.getRecentBlockhash();
+    transaction.recentBlockhash = recentBlockhash.blockhash;
+
     // Keypair signers.
     if (keypairs.length > 0) {
       transaction.partialSign(...keypairs);