+++ packages/js/test/helpers/setup.ts
@@ -23,6 +23,7 @@ import {
 } from '@mithraic-labs/tokenized-euros';
 import * as psyoptionsAmerican from '@mithraic-labs/psy-american';
 import * as spl from '@solana/spl-token';
+//@ts-ignore
 import { OptionMarketWithKey } from '@mithraic-labs/psy-american';
 import { bignum } from '@convergence-rfq/beet';
 import { Pyth } from '../../../../programs/pseudo_pyth_idl';
@@ -39,6 +40,8 @@ import {
   token,
   walletAdapterIdentity,
   Signer,
+  //@ts-ignore
+  createAmericanProgram,
 } from '@/index';
 const { mintOptions } = instructions;
 import { makeConfirmOptionsFinalizedOnMainnet } from '@/types';
@@ -282,7 +285,7 @@ export const createPriceFeed = async (
   return collateralTokenFeed.publicKey;
 };
 
-export const createAccountsAndMintOptions = async (
+export const createEuroAccountsAndMintOptions = async (
   convergence: Convergence,
   euroMeta: EuroMeta,
   euroMetaKey: PublicKey,
@@ -410,50 +413,30 @@ export const createAccountsAndMintOptions = async (
 
   await txBuilder.sendAndConfirm(convergence, confirmOptions);
 };
-
+//@ts-ignore
 const psyOptionsAmericanLocalNetProgramId = new anchor.web3.PublicKey(
   'R2y9ip6mxmWUj4pt54jP2hz2dgvMozy9VTSwMWE7evs'
 );
 
-export const initializePsyoptionsAmerican = async (
+export const createAmericanAccountsAndMintOptions = async (
   convergence: Convergence,
+  americanProgram: any,
   underlyingMint: Mint,
-  quoteMint: Mint,
-  taker: Signer,
-  maker: Signer,
-  quoteAmountPerContract: anchor.BN,
-  underlyingAmountPerContract: anchor.BN,
-  expiresIn: number
+  optionMarket: any,
+  // optionMarketKey: PublicKey,
+  optionMintKey: PublicKey,
+  writerMintKey: PublicKey
 ) => {
-  // const payer = taker;
-
-  const provider = new anchor.AnchorProvider(
-    convergence.connection,
-    new anchor.Wallet(taker as Keypair),
-    {}
+  const maker = Keypair.fromSecretKey(
+    new Uint8Array(
+      JSON.parse(readFileSync('./test/fixtures/maker.json', 'utf8'))
+    )
   );
-  anchor.setProvider(provider);
-
-  const AmericanProgram = psyoptionsAmerican.createProgram(
-    psyOptionsAmericanLocalNetProgramId,
-    provider
+  const taker = Keypair.fromSecretKey(
+    new Uint8Array(
+      JSON.parse(readFileSync('./test/fixtures/taker.json', 'utf8'))
+    )
   );
-
-  const expiration = new anchor.BN(Date.now() / 1_000 + expiresIn);
-
-  const { optionMarketKey, optionMintKey, writerMintKey } =
-    await psyoptionsAmerican.instructions.initializeMarket(AmericanProgram, {
-      expirationUnixTimestamp: expiration,
-      quoteAmountPerContract,
-      quoteMint: quoteMint.address,
-      underlyingAmountPerContract,
-      underlyingMint: underlyingMint.address,
-    });
-
-  const op = (await psyoptionsAmerican.getOptionByKey(
-    AmericanProgram,
-    optionMarketKey
-  )) as OptionMarketWithKey;
 
   const takerOptionToken = await spl.createAssociatedTokenAccount(
     convergence.connection,
@@ -487,44 +470,137 @@ export const initializePsyoptionsAmerican = async (
     taker.publicKey
   );
 
-  // const makerUnderlyingToken = await spl.getAssociatedTokenAddress(
-  //   underlyingMint.address,
-  //   maker.publicKey
-  // );
-
-  const optionMint = await convergence
-    .tokens()
-    .findMintByAddress({ address: optionMintKey });
-
-  const ix1 = await psyoptionsAmerican.instructions.mintOptionV2Instruction(
-    AmericanProgram,
+  const ixs = await psyoptionsAmerican.instructions.mintOptionV2Instruction(
+    americanProgram,
     takerOptionToken,
     takerWriterToken,
     takerUnderlyingToken,
     new anchor.BN(10),
-    op as psyoptionsAmerican.OptionMarketWithKey
+    optionMarket as psyoptionsAmerican.OptionMarketWithKey
   );
-  ix1.signers.push(taker as Keypair);
-  const tnx1 = new anchor.web3.Transaction();
-  tnx1.add(ix1.ix);
-  await provider.sendAndConfirm(tnx1, ix1.signers);
-  // const ix2 = await psyoptionsAmerican.instructions.mintOptionV2Instruction(
+  const ix1 = ixs.ix;
+
+  ixs.signers.push(taker as Keypair);
+
+  ixs.ix.keys[0] = {
+    pubkey: taker.publicKey,
+    isSigner: true,
+    isWritable: false,
+  };
+
+  const txBuilder = TransactionBuilder.make().setFeePayer(
+    convergence.rpc().getDefaultFeePayer()
+  );
+
+  txBuilder.add({
+    instruction: ix1,
+    signers: ixs.signers,
+  });
+
+  const confirmOptions = makeConfirmOptionsFinalizedOnMainnet(convergence);
+
+  await txBuilder.sendAndConfirm(convergence, confirmOptions);
+};
+
+export const initializePsyoptionsAmerican = async (
+  convergence: Convergence,
+  // underlyingMint: Mint,
+  quoteMint: Mint,
+  taker: Signer,
+  maker: Signer,
+  quoteAmountPerContract: anchor.BN,
+  underlyingAmountPerContract: anchor.BN,
+  expiresIn: number
+) => {
+  // const payer = taker;
+  // const provider = new anchor.AnchorProvider(
+  //   convergence.connection,
+  //   new anchor.Wallet(taker as Keypair),
+  //   {}
+  // );
+  // anchor.setProvider(provider);
+  // const AmericanProgram = psyoptionsAmerican.createProgram(
+  //   psyOptionsAmericanLocalNetProgramId,
+  //   provider
+  // );
+  // const americanProgram = createAmericanProgram(convergence);
+  // const expiration = new anchor.BN(Date.now() / 1_000 + expiresIn);
+  // const { optionMarketKey, optionMintKey, writerMintKey } =
+  //   await psyoptionsAmerican.instructions.initializeMarket(americanProgram, {
+  //     expirationUnixTimestamp: expiration,
+  //     quoteAmountPerContract,
+  //     quoteMint: quoteMint.address,
+  //     underlyingAmountPerContract,
+  //     underlyingMint: underlyingMint.address,
+  //   });
+  // const op = (await psyoptionsAmerican.getOptionByKey(
+  //   americanProgram,
+  //   optionMarketKey
+  // )) as OptionMarketWithKey;
+  // const takerOptionToken = await spl.createAssociatedTokenAccount(
+  //   convergence.connection,
+  //   taker as Keypair,
+  //   optionMintKey,
+  //   taker.publicKey
+  // );
+  // await spl.createAssociatedTokenAccount(
+  //   convergence.connection,
+  //   maker as Keypair,
+  //   optionMintKey,
+  //   maker.publicKey
+  // );
+  // await spl.createAssociatedTokenAccount(
+  //   convergence.connection,
+  //   maker as Keypair,
+  //   writerMintKey,
+  //   maker.publicKey
+  // );
+  // const takerWriterToken = await spl.createAssociatedTokenAccount(
+  //   convergence.connection,
+  //   taker as Keypair,
+  //   writerMintKey,
+  //   taker.publicKey
+  // );
+  // const takerUnderlyingToken = await spl.getAssociatedTokenAddress(
+  //   underlyingMint.address,
+  //   taker.publicKey
+  // );
+  // const makerUnderlyingToken = await spl.getAssociatedTokenAddress(
+  //   underlyingMint.address,
+  //   maker.publicKey
+  // );
+  // const optionMint = await convergence
+  //   .tokens()
+  //   .findMintByAddress({ address: optionMintKey });
+  // const ixs = await psyoptionsAmerican.instructions.mintOptionV2Instruction(
   //   AmericanProgram,
-  //   makerOptionToken,
-  //   makerWriterToken,
-  //   makerUnderlyingToken,
+  //   takerOptionToken,
+  //   takerWriterToken,
+  //   takerUnderlyingToken,
   //   new anchor.BN(10),
   //   op as psyoptionsAmerican.OptionMarketWithKey
   // );
-  // ix2.signers.push(maker as Keypair);
-  // const tnx2 = new anchor.web3.Transaction();
-  // tnx2.add(ix2.ix);
-  // await provider.sendAndConfirm(tnx2, ix2.signers);
-  return {
-    op,
-    optionMarketKey,
-    optionMint,
-  };
+  // const ix1 = ixs.ix;
+  // ixs.signers.push(taker as Keypair);
+  // ixs.ix.keys[0] = {
+  //   pubkey: taker.publicKey,
+  //   isSigner: true,
+  //   isWritable: false,
+  // };
+  // const txBuilder = TransactionBuilder.make().setFeePayer(
+  //   convergence.rpc().getDefaultFeePayer()
+  // );
+  // txBuilder.add({
+  //   instruction: ix1,
+  //   signers: ixs.signers,
+  // });
+  // const confirmOptions = makeConfirmOptionsFinalizedOnMainnet(convergence);
+  // await txBuilder.sendAndConfirm(convergence, confirmOptions);
+  // return {
+  //   op,
+  //   optionMarketKey,
+  //   optionMint,
+  // };
 };
 
 export const assertInitRiskEngineConfig = (
@@ -568,3 +644,23 @@ export const assertInitRiskEngineConfig = (
     model: 'config',
   });
 };
+
+// export const x = (
+//   provider: anchor.AnchorProvider,
+//   convergence: Convergence
+// ) => {
+//   const psyOptionsAmericanLocalNetProgramId = new anchor.web3.PublicKey(
+//     'R2y9ip6mxmWUj4pt54jP2hz2dgvMozy9VTSwMWE7evs'
+//   );
+//   // const provider = new anchor.AnchorProvider(
+//   //   convergence.connection,
+//   //   new anchor.Wallet(convergence.rpc().getDefaultFeePayer() as Keypair),
+//   //   {}
+//   // );
+//   // anchor.setProvider(provider);
+
+//   return psyoptionsAmerican.createProgram(
+//     psyOptionsAmericanLocalNetProgramId,
+//     provider
+//   );
+// };